remove dependencies for target; use docker-compose CLI instead of docker
The `seed-api-with-sample-data` make target does not depend on other
tasks anymore: it wouldn't make sense to force a clean slate to seed
data, and in the previous form the dependencies would not even work
correctly because `docker-compose up` (in the `start` target) executes
in the foreground - and detaching it would not help either because we
would need to know when the `postgresql-api` service is up before trying
to create data.

Moreover, the previous `docker` CLI invocation would not work either
because of TTY handling: by switching to a `docker-compose` CLI
invocation we keep the CLI interface consistent *and* this works as
expected.