new migration: add missing cascades
In general, join tables (such as `users_scenarios`) should not block
deletion of entities they reference. For example, if a user is deleted,
their roles in projects or organizations should be deleted too.

Some of these cascades were missing in the initial table definitions and
are added in the migration in this commit.