purge all expired tokens from db
We do this every time we issue a new token. In theory, this should be
done via a distinct API endpoint with admin role guard, which should be
triggered via a cron task _somewhere_.

In practice, doing this the way it's set up in this commit should not
add much latency to login requests: on very busy systems where millions
of users are active every day (thinking grandiosely here), there would
me many issued tokens at any given time, but these would also be purged
very frequently, every time a user logs in or refreshes their tokens, so
the delta should be fairly constant.

Of course rogue actors could abuse this (no matter how busy an instance
is) and repeatedly request JWT tokens (for example, using pwned
credentials), which may result on long-ish purge operations, but this
should be dealt with through rate limiting at a WAF or other layer in
front of the API.