make sure we reject payloads that contain non-whitelisted props
I don't see much of a compelling case for not being strict here,
especially in a public API. A DTO is a DTO, and we don't want to allow
properties that are not whitelisted (that is, are decorated with at
least one class-validator decorator - at least `@Allow` if no other
decorator is applicable to a specific property).

Not rejecting payloads with non-whitelisted properties would, among
other things, require to reject at the service layer requests that
contain an `id` property. We would not normally include id properties in
create/update DTOs, but without the strict check added in this commit we
would be accepting payloads with ids, while resorting to some typing
hacks to even be able to check for the presence of such props in
payloads (because the DTO would say there is no such property in the
type!).