automate running of e2e tests
Here we finally get to run all the tests in an isolated environment
(ad-hoc dbs with no persisted data).

* start from clean slate (in case anything was left around from previous
  interrupted runs)
* build images
* run migrations
* load test data (api and geoprocessing services)
* run tests
* teardown

The setup in this commit is super hacky - basically I have copied over
the two whole seed-api-with-test-data and seed-geoapi-with-test-data
targets replacing all the docker-compose config from these with the e2e
one.

We must clean this up soon, but this is a Friday evening and a cut
corner may be excusable as a tradeoff for starting next week with a
working test pipeline ready to be set up in CI, replacing the current
one which does not include loading test data.