refactor UserModule e2e tests; add tests for MARXAN-184
Improve separation of test sets to simplify how we handle authentication
across password resets, account deletions, etc.

This commit also adds tests for MARXAN-184: users should be able to
create a new account using the same email address as that which used to
be associated to an account that has been deleted.