(wip) add constraint on which rows are updated
Without this constraint, all rows (of any scenarios in any project)
would be overwritten.

For rows of the scenario being processed, the value of `protected_area`
would be set to the correct value; for all other rows (other scenarios
in any project), `protected_area` would be silently set to null as there
is no data coming in via the CTE for rows that don't belong to the
scenario being processed.