use gen_random_uuid() avoiding uuid-ossp
As we only need UUIDv4 and nothing else that uuid-ossp provides,
there's actually no need to rely on `uuid-ossp`.

In practice this change came about as an attempt to address an issue
with a constraint violation in a postgres system catalog table
(see:
https://github.com/Vizzuality/marxan-cloud/pull/614#issuecomment-968638969),
which however at closer look is unrelated to the `uuid-ossp` extension.

Regardless, this change removes an avoidable dependency.