run each migration in its own transaction
Running the whole set of migrations in a single transaction (the TypeORM
default) would cause issues with some enums not being committed before
they are used. This wasn't causing issues so far in the CI pipelines
(where the change in this commit is relevant) just because of an
unrelated bug that caused migrations runs to be tried twice (with
nothing being done the second time through the shell command changed in
this commit as the migrations had already run).