use hash of the_geom as proxy of the_geom in index
The previous implementation of the
`planning_units_geom_the_geom_type_project_id_check` index would not be
ok for `the_geom` data rows past the current limit for btrees (typical
errors: `index row size <NNN> exceeds btree version 4 maximum 2704 for
index "planning_units_geom_from_shapefile_the_geom_type_coalesce_idx")`,
so we are indexing a hash of `the_geom` here as a proxy of the content of
`the_geom` proper.

md5 should be faster than alternative hashing functions such as sha2xx,
while functionally enough for the purpose: the uniqueness check here is
meant to avoid *routine* duplication of planning units, and we are not
considering this as part of a threat model where a malicious actor would
purposely want to forge collisions.

No realistic chance of collision is expected through normal use of the
platform, and even if this were to happen, the outcome would be some
wasted storage space for *a specific* planning unit, therefore making
this computationally unattractive even for an hypothetic malicious actor.