temporary workaround for error when loading bioregional features
Checking whether `properties` is defined here is just a
workaround to avoid an error when processing `bioregional`
features, which would prevent progressing through the stop of
configuring features for a scenario until this code is reviewed.
Without much knowledge of the flow for feature data, I see that
short-circuiting the `map()` below and therefore setting
`splitOptions = []` still results in properties being shown in the
dropdowns used for splitting features, but since `properties` is
always undefined (from what I can see), we may need to adapt the
API payload or how we process it here.