upsert when users try to upload an identical custom protected area
The initial implementation was rejecting uploads of custom protected
areas where the geometry exactly matched an existing one for the same
project. However, this would result in an application failure that had
no recovery path: this could be added, however the change in this commit
simply switches to an upsert so that a duplicated upload is handled
gracefully (and can be actually used to upload the display name of the
protected area network being uploaded).