return as right (and add typing)
This commit should fix a recent regression, whereby the missing
`right()` wrapping of the return value from
`ScenariosService.uploadLockInShapeFile()` caused the response payload
to not match what the app would expect (though the app would just
silently fail, hanging indefinitely).

The regression may have slipped through because of the lack of typing on
the response from the Axios call through which we request the GeoJSON
version of the shapefile that we submit to the geoprocessing service:
this change also adds strict typing (and removes properties from
features, since these are not needed in this context).