forcibly release all locks active at the time of migration
This is so we don't risk running into a constraint violation when adding
a FK reference later in the migration.

There should really be no "active" users by the time this migration
runs, as there is no production environment yet, and LOFU should help
the few developers and test users to keep their sessions running
smoothly nevertheless.

I am not sure how wise it is to use an Unicode emoji in a PostgreSQL
comment, but apparently we'll be fine with this 😇.