don't bother setting a grid shape if we have no GADM ids
This is now going to trigger an error thanks to new validators. In
practice, no project should be created without some kind of grid setting
(either using a custom grid or a generated hex or square grid), but
since many tests rely on doing this, I am not going to change this for
the time being (especially as the frontend app doesn't attempt to do
so).