update and enforce typing via assertion
This was one of the leftover type issues in the NestJS v7->v8
transition, albeit not related to NestJS itself but to
BullMQ/Typescript. As we should in practice always receive progress data
as an object in either of the shapes expected under the ProgressData
union type, we add a type assertion here, and throw an error if we
receive unexpected data.