(delete) add api config items to geoprocessing service
This is a hack to allow the geoprocessing service to start again in k8s
environments.

Recent changes (which I could not identify with certainty) started
causing more parts of the API code to be transpiled when Docker images
for k8s deployments are built, and included in the built `dist` folder.

This would in turn lead to some of the API's own config items to be
looked up when starting up the geoprocessing service (unsuccessfully,
since they are not available in the config setup for the geoprocessing
service).

TL;DR we may have had a few direct dependencies between libs and api or
geoprocessing and api for a while and never noticed since everything
(both services and all the libs) is built and shipped in Docker images
run via Compose, but once the builds started including in the
geoprocessing service intertwined API code that references config items,
the stricter separation of environments in k8s led to failure to start
the geoprocessing service.

This commit temporarily works around this issue by exposing all the API
config items in the geoprocessing service, and providing the same
defaults for these that would be used in the api service.

The proper solution is of course to undo these dependencies and then
undo all the changes in this commit. Since this will not be
straightforward though, in the meanwhile the hack in this commit should
allow the geoprocessing service to at least start again.