remove allowedHeaders whitelisting
This was an historical artifact inherited from earlier NestJS setups,
but in practice, I don't think we need to be so picky with request
headers as long as we whitelist origins correctly, as we seem to be
doing.

Somehow, the previous setup was working correctly for the frontend app,
except when this is puppeteered via the webshot service.