use in-container port by default
I think we were relying on a networking configuration that is either
valid in Compose v1 and not in v2 or on some combination of factors that
made it possible in Compose v1 to reach a service running in a
container, from a container on the same Docker network, using a port
exposed via the `services.<servicename>.ports` configuration section,
whereas this is not possible (arguably, for very good reasons) in
Compose v2.

This commit fixes the `env.default` file so that it can be used "out of
the box" after running `make .env` in a development instance.